Focused primarily on Arch and Lib documntation.
With this commit, I realize I've deviated from the stated purpose of this branch. While there is a LOT of Calypso-focused exploration (NONE of it worth saving), on Friday the demonstration I gave to the MQTT team about how I used Doxygen at another company, prompted me to begin the documentation on this project that I've been putting off for quite some time.