End of home work on Board support.
A bit of documentation. This edition "works" with 3 simulated button
presses, in MinGW/Eclipse, with the caveat that simultaneous simulated
button events are not handled (only the last one is handled).

You can get around this by arranging the simulation input stream for at
least a one-sample offset between events. The justification is this:
* It's only a unit test, and the code that is "weak" is part of the UT
environment, not the reusable library code
* The chances of truly simultaneous button-press events in real life is
practically nil.
Ergo, it's good enough for now.