End of day commit from personal machine. Nothing works. Event Queue.
In the midst of copying from my throw-away prototype, not in this repo, to my official repo, simply ran out of time for the day and won't get a chance to get back on it before the weekend. Since I'm not bringing my home machine to my weekend domicile, I need to save it as-is.